pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jira-comment-assistant"
version = "0.1.0"
description = "AI assistant for responding to developer comments on Jira defects with context-aware drafts"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Nagasai", email = "nagasai.chintalapati@hpe.com"},
    {name = "Yousef", email = "yousef.konswah@hpe.com"}
]

dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "atlassian-python-api>=3.41.0",
    "pydantic>=2.0.0",
    "github-copilot-sdk>=0.1.23",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.27.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
